+++ packages/common/src/index.ts
@@ -2,3 +2,5 @@ import 'reflect-metadata';
 export * from './constants.js';
 export * from './interfaces/index.js';
 export * from './decorators/index.js';
+export * from './decorators/rate-limit.decorator.js';
+export * from './errors/shardix-error.js';