+++ packages/cache/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @shardix/cache
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @shardix/common@0.4.0
+
 ## 0.1.2
 
 ### Patch Changes