+++ packages/config/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @shardix/config
 
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @shardix/common@0.4.0
+
 ## 0.2.2
 
 ### Patch Changes