+++ packages/discordeno/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @shardix/discordeno
 
+## 0.5.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @shardix/common@0.4.0
+
 ## 0.5.0
 
 ### Minor Changes