+++ packages/discordjs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @shardix/discordjs
 
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @shardix/common@0.4.0
+
 ## 0.2.0
 
 ### Minor Changes