+++ packages/provider-dashboard/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @shardix/provider-dashboard
 
+## 0.6.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @shardix/common@0.4.0
+  - @shardix/dashboard-api@0.5.2
+
 ## 0.6.1
 
 ### Patch Changes