+++ packages/provider-queue/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @shardix/provider-queue
 
+## 0.4.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @shardix/common@0.4.0
+
 ## 0.4.2
 
 ### Patch Changes