+++ packages/runtime-distributed/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @shardix/runtime-distributed
 
+## 0.4.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @shardix/common@0.4.0
+  - @shardix/core@0.4.0
+  - @shardix/cluster@0.5.2
+  - @shardix/ipc@0.4.3
+
 ## 0.4.2
 
 ### Patch Changes